boottest: Correctly use file:/// schemes
Wikipedia says it best: "while "file:///foo.txt" is valid,
"file://foo.txt" is not, although some interpreters manage to handle the
latter".

File schemes must be absolute and expressed with three slashes '/'. It
seems url.Parse() and url.String() make this assumption at times, which
makes it difficult to modify the grub interpreter to continue emitting
incorrect URLs with two slashes.

This change permits the grub interpreter to emit file schemes with
absolute paths. The boottest package takes these paths and strips out
the non-hermetic portion in before comparing against the value in the
json.

Signed-off-by: Ryan O'Leary <ryanoleary@google.com>