securelaunch: Read the hash of the public key from the TPM and verify it
We verify the policy file by using a public key and signature, however we
currently do not verify the public key file in any way. An attacker could
cause an unwanted policy file to be used by also replacing the public key
and signature files.

Instead, read a stored copy of the hash of the public key from the TPM
and use that to verify that the public key file is the expected file.
This lets us trust it and thus trust that the policy file hasn't been
tampered with (assuming it verifies OK).

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>