disk_unlock should try unique HSS and allow retries
In the normal case there are multiple identical copies of the HSS, and
trying the same HSS multiple times would consume password retry
attempts on the drive, which most drives have a limited number of.
Instead we only try unique HSS values.

We also introduce the option to have multiple retries for an HSS in case
there's failure for any reason other than an incorrect password.

Signed-off-by: pallaud <plaud@google.com>