Generalized checker framework
The checker framework is meant to help with troubleshooting and
auto-remediations when a host does not boot. The original code required
a hardcoded checklist at compile time. This diff expands the framework
to parse a JSON file containing the checklist, to allow for more
flexibility. For example, it is now easier to run platform-specific
checklists by just embedding platform-specific JSON configurations,
rather than hardcoding all the checks and known platforms.

This change also allows any function that returns an `error` to be used
for checks. In order to do that, reflection is used to check the
function signature, allow any parameter type (instead of, for example,
just strings), and to create a callable function with the appropriate
parameters.
This makes its usage easier and more immediate, as long as the function
can return an error. From an usage and implementation perspective, it is
very similar to the `FuncMap` type in Go's standard library's
`text/template` package, see
https://golang.org/pkg/text/template/#FuncMap .

It is also possible to use Go templates in the JSON file. In fact, every
JSON containing a checklist is executed as a template, and the
parameters can be passed using the exampole command line (parameter `-p`
under `examples/checker`).
The example command line also provides a subcommand to print a sample
configuration.

Signed-off-by: Andrea Barberio <insomniac@slackware.it>