provide an option to use shellbang, not symlinks, for busybox
Not all kernels support symlinks, and not all file systems
on all kernels do. Symlinks are not as portable as we want.
For example, a simple u-root on a VFAT file system can not
be created.

It turns out there is a way to use shellbang files that is as
efficient as symlinks and much more portable.

Add an option to the busybox builder to generate shellbang files,
not symlinks.

TODO: update the tests. But this change is known to work on Plan 9.

Question: should we just move to shellbang files
only, given their more universal nature, or keep it as an option?

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>