ash: make it work on Plan 9
This is how we learn to make elvish work on Plan 9.

This is actually sort of working with minimal complete.

Issues:
our 9p server only returns one entry at a time on a directory read,
which makes it horrendously slow to complete. I.e. it works like
9p2000.L or Unix in general. Too bad.

hmm, that's all the issues. The completion would be pretty fast
were it not for that.

It's another reason shells tend to cache everything in $PATH, actually.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>