mount: with no args, show mounts
In the early days of u-root we had hoped to cleave to a more
plan 9-like model, in which the operations of mounting were
two separate commands. But the habit of typing mount by itself
to see mounts is not going to go away.

When mount is invoked with no args, it will try to read the mtab
from a number of standard places (Linux, Unix, OSX) or from
/proc/os.Getpid()/ns (Plan 9).

As an experiment, to try to keep the number of system calls down,
handle the "no args" case in init. We still end up with 300, which
is largish.

While we are here, fix a bug in the Plan 9 version, and
remove its use of panic()

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>