uroot: recurse ldd directories and stop trying to ldd non-ELF
Users frequently wish to include whole directories of executables.
Further, messages about ldd parsing are confusing when the
LDD on a non-regular-file fails.

Change the enumeration of the single file to a filepath.Walk.
Only try to enumerate ldd on files that are ELF files.

This change makes the use of -files a much better experience,
as well as allowing users to include whole directories
containing executables.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>