pkg/cpio, cmds/core/cpio: Allow reading from pipes
Currently, the cpio package can not read from pipes. This violates
a fundamental property of both cpio and Unix in general.

The fix might seem to be allowing creation of Readers from an os.File.
We could then see if the File in question implements Reader or
ReaderAt: if ok, rat := f.(ReaderAt) ...
and handle each case, transparent to the caller.

The problem is that some Go os.File types, such as Pipes,
advertise themselves as implementing ReaderAt but in fact do not
do so.

e.g.
	r, _, _ := os.Pipe()
	var b [2]byte
	var _ io.ReaderAt = r
	_, err := r.ReadAt(b[:], 0)
	fmt.Printf("%v", err)

will fail, since whatever Go may think, Linux pipes do not allow a
seek, even when the offset is not changing. So the test
above will not work.

The Go runtime needs to be fixed, so that it does not claim
pipes implement io.ReaderAt, but it's going to be a long
time before such a change propagates everywhere. The cpio package
must accomodate this error. Further, there will almost
certainly be other future types which make this mistake.

Provide a NewFileReader function, which accepts *os.File
and returns a Reader. The function performs a simple test,
io.Seek(0,0), and, if that works, just creates a Reader
using the File's ReadAt.

If that seek fails, the function wraps the file in a
discardreader.

It is best to use a ReadAt when possible to avoid having
to read and discard data to implement seek.

cpio can now use stdin from a pipe.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>