pkg/io/ports_linux: only allow outb and inb
/dev/port on linux only does an outb/inb.
see drivers/char/mem.c::write_port and read_port.

This can cause real trouble for hardware. For example,
on an AMD Rome system, writing to 0xcf8 with a write
system call to /dev/ports hangs the system in a way the
BMC can not recover from. This is because the 4-byte write
is realized as four outb instructions, not one outl.

Hence, on Linux, the memio package should return an error
if anything other than a uint8 is written to /dev/port.

Note that this restriction does not apply to Plan 9,
which has a better design for port IO than Linux.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>