Add Jessie Frazelle's container package
It has not been maintained for 3 years but has a lot of utility,
include the ability to self-sandbox. The thought is to optionally
incorporate this into our busybox.

I've talked to Jessie and she is OK with our incorporating this
code, in hopes it gets wider use.

This includes another license check, for MIT licenses.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>