tar: Support non-empty files with 0 size
For example, this includes most of the files in /proc and /sys under
Linux. The size cannot be found with a stat. No, you have to read the
whole file first to determine size. Since we cannot seek in a pipe (such
as when piping tar into gzip), this reads the file into a memory buffer.

This commit also adds the --no-recurison option at the same time. The
new unit test needs it to avoid taring all of /proc.

Signed-off-by: Ryan O'Leary <ryanoleary@google.com>