sluinit: Don't ExitOnError for flag.Parse()
By default, if flag processing for the command line hits an error, it
will exit. However, this has the side-effect in u-root of dropping it
to the u-root shell. This is undesireable behaviour, as it makes it
trivial for a malicious user to gain u-root shell access.

Change the default error handling behaviour for u-root command line
parsing to be ContinueOnError so the u-root code can handle any
errors, or at least not drop to a shell.

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>