mount: use magic numbers for mounting
The method of TryMount trying each file system type in turn to mount is
causing lots of confusion. The kernel is unbelievably noisy on these
failures and it is very hard to tell errors from noise.

This change makes TryMount smarter. It opens the device,
ensures we can read at least 64k, tries to find the magic,
and, if all that works, tries the mount. This path is only taken
if the user has not specified the file system type.

Some file systems have the same magic number (ext2, 3, 4). For these
file systems, the code will try to find the "highest priority" one supported
in the kernel, and use that. E.g., if a device has the magic for ext2, 3, or 4,
but only ext3 is in the kernel, the code will try to mount as ext3.

Reading 64K is a compromise; 1M is preferred but on SPI parts
that is very slow.

Some file systems can never be mounted writeable, and the code
further provides flags to be added to the mount flags so an error
message can be avoided. Setting additional flags is supported.

One failing is the code can not accomodate a sparse magic, i.e.
one like FAT32 where the first two bytes and then some later bytes
indicate it is vfat. That's kind of ok; as far as is known vfat can mount
fat16, and mkfs.fat does not like to create fat16 anyway.

Only a small number of file systems are filled in at present: ext*, vfat, xfs,
iso9660. But that small number gets most people very far.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>