pox: Add self-executing pox archive
The generated archive can now execute itself:

    $ pox -cvsf date /bin/date
    Creates a self-executing pox called "date".
    $ ./date --utc

The archive is simultaneously an elf file and zip file:

    $ file date
    date: ELF 64-bit LSB executable, x86-64, version 1 (SYSV), statically linked
    $ unzip -z date
    Archive:  date
    /bin/date
      Length      Date    Time    Name
    ---------  ---------- -----   ----
            0  2021-03-02 08:07   /bin/
       113864  2021-03-02 08:07   /bin/date
            0  2021-03-02 08:07   /dev/
            0  2021-03-02 08:07   /lib/
	    ... elided ...
    ---------                     -------
      4149104                     12 files

The trick is zip files place their header at the end of the file, while
elf files place theirs at the beginning. The zip file has a special
piece of metadata called the "comment" we use to store the path to the
binary to execute by default.

Signed-off-by: Ryan O'Leary <ryanoleary@google.com>