testramfs: name the mount "source"
testramfs uses a tmpfs mount, i.e. there is no source.
For the tmpfs mount it used the "empty source" in the Mount system call:

if err := syscall.Mount("", tempDir, "tmpfs", 0, ""); err != nil {

While this is perfectly legal, and was even common, it leads nowadays to odd-looking
mountpoints:

cat /proc/mounts
...
 /tmp/u-root621209272 tmpfs rw,relatime 0 0

Note the leading ' ', which is due to the 'empty source'.

This leads to the following non-intuitive message from snap on ubuntu systems:
rminnich@a300:~/oreboot$ emacs src/mainboard/amd/romecrb/Makefile*
2021/04/09 09:15:58.634892 system_key.go:129: cannot determine nfs usage in generateSystemKey: cannot parse mountinfo: incorrect number of tail fields, expected 3 but found 2
2021/04/09 09:15:58.640801 cmd_run.go:994: WARNING: cannot create user data directory: cannot determine SELinux status: failed to obtain SELinux mount path: incorrect number of tail

ubuntu is unable to parse a valid /proc/mounts! But this problem is likely to be widespread,
so it pays to fix it with a simple change:

if err := syscall.Mount("testramfs.tmpfs", tempDir, "tmpfs", 0, ""); err != nil {

The mountpoint now looks like this:
testramfs.tmpfs on /tmp/u-root231497718 type tmpfs (rw,relatime)

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>