Fixups and cleanups
Add comprehensive constants.

Do not try to mmap the low 2G -- it would work fine but Linux
will not allow it any more. Implement with a new type,
offsetReader, which mmaps a file and returns an io.ReaderAt.

Properly distinguish "there is no coreboot table" and
"could not parse coreboot table"

Find coreboot console buffer, whereever it may be.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>