pkg/pci, cmds/core/pci: add more printing per requests from engineers
The original pci command was modeled after the Plan 9 version.
Hence the original u-root pci command was very simple. On Plan 9,
extra capabilities are added by extra commands. Kind of like Unix.

Simple is not the habit on Linux; the style of "one command that does everything
possible" is more the mode.

And it turns out users like that; they want more printing like lspci.

This is the first of series of changes to give users what they want.

For now, the command will show Control and Status registers for selected
devices when -v=1. The information presented is basically the same as lspci.
Note that for Status, INTx is presented in bit numbered order on pci, and lspci
for some reason puts it last. Your guess is as good as mine.

There is also, now, the ability to dump to JSON
for forensics later.

rminnich@t510:~/go/src/github.com/u-root/u-root/cmds/core/pci$ ./pci -v=1
0000:00:00.0: Intel Corporation Core Processor DRAM Controller
00000000  86 80 44 00 06 00 90 20  02 00 00 06 00 00 00 00  |..D.... ........|
00000010  00 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  |................|
00000020  00 00 00 00 00 00 00 00  00 00 00 00 aa 17 93 21  |...............!|
00000030  00 00 00 00 e0 00 00 00  00 00 00 00 00 00 00 00  |................|

	Control: I/O- Memory+ DMA+ Special- MemWINV- VGASnoop- ParErr- Stepping- SERR- FastB2B- DisInt-
	Status: INTx- Cap- 66MHz+ UDF- FastB2b- ParErr+ DEVSEL- DEVSEL=fast <MABORT- >SERR- <PERR+

etc. etc.

Note that this is the output presented for lspci -v -v; there are no plans to present the even simpler
output for lspci -v, absent demand. It is not really that useful as compared to -v -v.

The config space dumping will go back to only being enabled by -c in the next PR.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>