pkg/termios: fix bugs in sgtty parsing
There were a number of bugs, and error handling was not well done.

"rows" was mistyped as "row"
"cols" was mistyped as "col"
The darwin/unix differentiation was not well done at all (it's now about 10 lines)
The test did not test actual setting of values, which is why some bugs were missed.

The only difference anyone should see is that some commands work now.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>