u-root: Make it run from anywhere.
With the fix for go1.16, we lost the ability to run u-root from any
directory.

This change restores that capability and does some initial work for modules.
It relies on rsc's recent statement that GOPATH is never going away.

If
GO111MODULES=off
(increasingly rare and soon to be impossible for Go)
then the code behaves as it does today.

If GO111MODULES=on or is not set, the code will look for a go.mod.

In the discussion below, g=github.com, u=u-root,V=value of -version flag

The code first looks in the legacy location,
$GOPATH/s/g/u/u/go.mod and, if found, sets the modPath
variable (not yet used) to that path.

Else, it looks in
$GOPATH/go/pkg/mod/g/u/u+V
for go.mod, e.g. if version is @v1.0.0, the path
$GOPATH/go/pkg/mod/g/u/u@v1.0.0/go.mod
will be checked.

Note that one can set version to *, to find ANY matching go module,
or set it to "", to ensure that only the u-root tree in the legacy
location is used.

Finally, the code does an os.Chdir to filepath.Dir(modPath), so
it can once again be run from anywhere.

N.B.: I did experiment with golang.org/x/tools/go/packages but it's a bit
of a clumsy package to use, and does not quite work as advertised, so I went
with this simpler wired-in globbing. The modules path root in $GOPATH/pkg/mod
is unlikely to change at this point anyway, so it seems safe.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>