curl: add ways to fetch without cache.
Current Fetch for each scheme returns a uio.NewCachingReader
which buffers a copy of bytes as it reads. In cases where that
buffer is not needed.

e.g. On a machine with 4 GB ram and we need to download 1.1G netboot
image, and reader returned by Fetch can cause a OOM as we cache image
bytes read, but which is not re-used.

So we might as well provide FetchWithoutCache to give clients options
to fetch without the smart cache.

This is a no-op for places using Fetch.

Signed-off-by: David Hu <xuehaohu@google.com>