pkg/boot/menu: Add a generic return to caller
Add a generic return to caller menu option. This is added to
boot, pxeboot, and netbootxyz. If the booter was called from the
command line, it returns to the command line and command history,
etc is maintained. If the booter is called from systemboot it will
continue to the next booter in sequence.

Signed-off-by: Marc Jones <marcjones@sysproconsulting.com>