tools/makebb: add -c switch so it lists commands
In scripts, as used in, e.g., Harvey, it is important to
know what commands are built into a bb.

The glob passed to makebb will not always be the same
as the actual commands that get built (there might be a
match for a command that does not get built due to build
constraints).

The -c switch to bb will now list the commands that get built in.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>