rename linux-dependent commands / add !plan9 build tags
Linux dependencies started creeping in again.

The Plan 9 template will now include all of core
save for elvish (a lost cause if ever there was one ...)
to ensure we don't do this again.

We took this opportunity to clean up the rtc code, which did not
handle !linux builds correctly.

It would be nice to use the //go:build syntax but that does not pass our CI,
so the old form is used. For now.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>