pkg/bb: do not return an error if no main can be found
build constraints can result in there being no main, e.g.,
GOOS=plan9 u-root cmds/core/d*
should not be an error just because plan 9 does not have or want dmesg.

Why do this? Because linux-ism's keep creeping back into our code. The easiest way to
avoid them is to build everything with GOOS=plan9, and let build constraints elide
the ones that are not supported.

For that to work, u-root has to ignore packages that have no main after build
constraints are applied, not error out on them.

This is a bit of a hack until we can get the new gobusybox code in use.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>