kexec: accept a list of files to be passed to the next kernel
The -extra switch allows inclusion of extra files. The files should not
be huge, as they are pulled into an in-memory archive.

Example usage:
In this example, we are on the console of a node.
cd /
cp /etc/hosts .
cp /dev/kmsg .
kexec -extra 'kmsg hosts' -i initrd.cpio kernel

node reboots ...
~/# ls -l kmsg hosts
-r-------- root 0 24295 Aug 27 15:58 kmsg
-rw-r--r-- root 0 1211 Aug 27 15:58 hosts

The files are available. The main use of this option
at the moment is to preserve the kernel log buffer
for the next kernel.

This is a proof of concept; change at will.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>