linux kexec: skip kernel / initrd copying when source is a regular file.
...and under tmpfs.

One caveat discussed is original file may be opened
for writting, which kexec file load does not like.
This change does not proactively prevent that from
happening within linux.go, but now allow it to fail
in that case.

One thing I can do is write, in golang, lsof, and matching
/proc/*/fd/* linked file against original file name
and see if there is any currently opening it, and throw
an error to kexec load caller, but that is not giving
any advantage over let kexec file load propogating up
the error. One benefit being we can do sth custom when
that happens, say proactively purging / closing them.
That just lead us into another hole of mess. So I prefer
to "Leap Before You Look" here for simplicity.

Signed-off-by: David Hu <xuehaohu@google.com>