change u-root to optionally use gobusybox
We need u-root to work better with modules. Previous changes made that possible,
and this last step is a step toward having the u-root command itself work
well with modules.

Change u-root so it can use both the old style and the gobusybox package,
defaulting to the old style.

The u-root command is widely used, so removing/renaming the command is not
acceptable.

Instead, we add an option, gobusybox, which will cause u-root to use the
gobusybox package, and hence work with modules.

This will, in turn, let us start using github actions, which seem to really want
us to use modules.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>