retract v6.0.1+incompatible
Modules are hell on a good day. It's just impossible to see why
go mod tidy does what it does. In this case, in the cpu tree,
it insists on bringing in v6.0.1 -- which, as far as I know,
was never even tagged.

Let's see if this retract fixes go mod tidy's insistence on bringing
in v6.0.1 in the cpu command.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>