gbb: use correct env
both u-root and gobusybox have their own pkg/golang. The Environ
structure is mostly the same, so let's just copy from one to the other.
This will allow programmatic users of the uroot API to cross-compile,
not just command-line users, and adds build tag support (which are added
to the u-root Environ in our main).

Signed-off-by: Chris Koch <chrisko@google.com>