cmds/core/init: fix flag message
The usage message for init includes this:
"print all build commands"

This is a leftover from source mode, when we did in fact compile all commands
on startup. Those days are gone (thanks modules!) so fix the message
to reflect what it really does.
"Enable libinit debugging (includes showing commands that are run)"

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>