Use existing setup data for the next kernel
The existing boot param data is in sys/kernel/boot_params/data.
This is, hopefully, the original; there are two directories in
there, also containing boot data, so it is not currently clear which
is best to use. So we start with this one.

This change gets us to working e820 tables.

It further allows us to remove the jmp . in the to64 purgatory.

We can now kexec kernels, using kexec_load, without needing
anything from the C kexec-tools. Further, the code is dramatically
simpler and easier to read. The next step is to test ways to make
VGA continue to work across kexec.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>