Change gosh tabcomp default
The tabcompletion library used causes a panic if on framebuffer is there.
So if u-root is accessed via a serial connection it will trigger the
panic and cause the shell to exit which usually means killing init and
therefore a kernel panic. So with this change gosh will start without
tabcompletion and tell the user how to enable it afterwards.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>