kexec_load: have KexecLoad use same args as Load
The Load functions should match. Further, in linux.go,
Load functions are called with a *os.File, but KexecLoad
used io.Reader as formal parameters. The result was that when
the ramfs io.Reader was tested for != nil, the test was always
true, leading to all kinds of fun.

See when-nil-isnt-equal-to-nil for why.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>