boot/linux: write a new package comment
My goal with this was to write a package comment for people who aren't
familiar with kexec_load yet. IMO the history is a bit counterproductive
there: the choices that are mentioned in the purgatory discussion are
not relevant today anymore. It would never make sense to have a
purgatory today that copies code (and it hasn't made sense since 2005,
which is when the Linux kernel started copying the code for you).

Signed-off-by: Chris Koch <chrisko@google.com>