termios: make struct Termios contain unix.Termios, not *unix.Termios
The Termios type is intended to be a value, not a reference,
so that it can be modified or saved and reused. Dereferencing
a *Termios should yield a value that, when changed, does not
affect the original.

For example, the Raw() function returns a Termios intended
to be used later to reset the terminal state, and modifies
a copy.

As part of bc40a4c6c, "make termios less linux/unix-specific.",
a Termios struct was created containing the *unix.Termios that
had been passed around directly.

This broke semantics of functions like Raw and MakeRaw,
which depend on being able to return an unmodified struct value
while modifying a copy of the one passed in. Since the passed-in
*Termios contained a *unix.Termios, copying just mean we had two
Termios using the same *unix.Termios.

This change fixes a problem we observed in the cpu command, and includes
a test to avoid future regressions.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>