Disallow import paths and switch to file paths
This commit adds a flag to specify the filesystem path
of the u-root source so that templates and import
paths of u-root commands get resolved correctly. In
addition all other non-filesystem paths will cause an error.
As we fully deprecate the old BusyBox mode and switch over
to gobusybox completely, it's sufficient to only allow
filesystem paths here as those should be correctly handled
by gobusybox.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>