cpio: fix to allow a same file as source for multiple extra files
If a same file is specified as source for multiple extra files,
e.g. -files 'README.md:a.txt' -files 'README.md:b.txt', the
contents of the second and subsequent files, e.g. b.txt, are empty.
This is because `ReaderAt` is not initialized after the second time
if inode number & device of source are duplicated.

Signed-off-by: Nobuhiro MIKI <nob@bobuhiro11.net>