Introduce uroot-source flag
This adds the uroot-source flag which is used to find the u-root
source tree on the disk. As we disallowed import paths, we need to find
out where u-root is on the disk in case templates or other core commands
are needed. If the UROOT_SOURCE env is set, it will be checked and used
in case the flag was not provided.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>