sluinit: Clean up iSCSI code
Currently, iSCSI devices are always scanned for, even if it wasn't set
on the kernel command-line. Since the command-line arguments missing
returns an error, but it's not necessarily an actual error (e.g., the
system isn't booting from an iSCSI device), the return value is ignored.
If the system was supposed to boot from an iSCSI device but there was
another error, then it will continue on and likely fail in some other,
unexpected way.

Check for the necessary command-line arguments ('netboot' and
'rd.iscsi.initiator') and only scan for iSCSI devices if they are present.
Catch any error from scanning for iSCSI devices and print the error and
quit.

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>