sluinit: Call an exit() function instead of returning from main()
If an error is encountered, then main() (usually) prints a message and
then calls return, which drops to the u-root shell. Instead, call an
exit() function that cleans up, prints the error, and loops forever
trying to reboot the system.

While we're at it, clean up the error and debug messages a bit.

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>