kernel placement flag is honored.
default to look for a 2MB aligned range from lowest
hole as possible. That satisfy it when Image header
flag bit 3 is 0. But it means kernel is free to be
placed anywhere when it is 1, which we might set
to still place as low as possible, FOR NOW.

In case, we have a practical use case / reason to
actually place kernel code differently when bit 3
of flag is 1, we can come back and add a "if" check
to do it dfferently.

Signed-off-by: David Hu <xuehaohu@google.com>