securelaunch: Measure policy with other files, not when getting it
The policy file is measured automatically when it is parsed, but a user
may not want to perform measurements. Store the loaded policy file in a
byte slice so that it can be measured later, if desired. This will also
let us eventually separate out loading the policy file from parsing it.

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>