Add Github actions for testing, building and everything CircleCI does
This commit adds several workflows to test and build u-root.
Most steps are copied from CircleCI but ran on GitHub now.
This also adds stale issue and PR management via a Github
action that runs on a schedule.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>

Add Github actions for testing, building and everything CircleCI does

This commit adds several workflows to test and build u-root.
Most steps are copied from CircleCI but ran on GitHub now.
This also adds stale issue and PR management via a Github
action that runs on a schedule.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>