Refactor align pkg.
*Now w/ 4 functions: Up, Down, UpPage, DownPage.
*move it from pkg/boot/align to pkg/align.
*have all (v +/- mask) &^ mask based alignment
 re-use this pkg.

NOTE that alignment size need be a power of 2
for bit operation based alignment in this package
to work.

An alternative formula by calculating remainder,

Up(v, alignmentSize)

    v + alignmentSize - v % alignmentSize

Down(v, alignmentSize)

    v - v % alignmentSize

which does not require alignmentSize be a power
of 2. Modulo is slow in C/C++ as  a % b is equivalent
to a - (a/b), which involves binary multiplication.
We run 64bit mostly, so two 64bit number multiplication
can takes up to 64 intermediate additions. Modern processor
can do it with 6. Golang's modulo implementation
a bit faster, but not as much as direct bit operation.

The bit operation based approach is what kernel
code uses, with alignment as a power of 2.

I think we can leave align package be as is for
aligning towards a power of 2.

But if we can prove we don't need the effiency of
bit operation, using modulo approach can help unify
all alignment code into align package.

Signed-off-by: David Hu <xuehaohu@google.com>