securelaunch: Dump out all PCR values at boot
In order to seal secrets in the TPM that only the securelaunch kernel can
unseal, we need to know the values of the PCRs while in the securelaunch
kernel. Dump out all the PCR values when starting u-root so they can be
used later to provision the system and seal secrets in the TPM.

This will be useful for future functionality. For example, storing the hash
of a public key file used to verify the securelaunch policy file or storing
a LUKS key that can be injected into the target initrd to decrypt disks
automatically without leaking the key (or even needing to know what the
key is).

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>