securelaunch: Use a public key and signature to verify the policy file
Currently, the policy file provided is trusted implicitly. However, an
attacker could simply swap the policy file out with another and it will
be processed.

The first step in addressing this issue is to verify the policy file
using a public key and signature. Currently, the key and signature are
both trusted and the attacker could simply switch out all the files.
A future commit will use a hash of the public key file, stored in the
TPM, to verify that the public key provided is the expected one.

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>