securelaunch: Add function to read a file, other clean ups
Add a function to read a file (mounting the containing device if
necessary). While we're at it, clean up the code for writing a file as
well. Currently, WriteToFile() takes a default file name and expects
the target location to be mounted. Instead, create WriteFile() that
will mount the device if needed and write the specified file. Modify
persist() (currently the only user of WriteToFile()) to handle the
case where we might need a default file name (as that's the only one
that cares about that).

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>