securelaunch: Only look for the policy file based on 'securelaunch_policy' flag
Currently, if the user doesn't specify the 'sl_policy' kernel command
line argument or if it can't load what the user specified, u-root scans
all the block devices looking in '/', '/efi', and '/boot' for a
securelaunch.policy file. This makes it too easy to load the wrong policy
file. Instead, just use whatever was supplied on the kernel command line.
If the flag isn't set or the policy file cannot be located or read,
return an error.

This is also the first step in preparing for verification of the policy
file by using a using a public-private key pair to sign the policy file.

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>