securelaunch: Refactor policy code
The policy code used to locate, read, measure, and parse the policy file
all in one function (named Get()). However, as we separate out code (e.g.,
measuring the policy file) and only "locate" it from a kernel command-line
argument, it makes less sense to keep the rest of the functionality
bundled. This will be even more true once verifying the policy file is
also added.

Split the functionality of the Get() function into Load(), Parse(), and
Measure() functions.

Signed-off-by: Patrick Colp <patrick.colp@oracle.com>