u-root.go: set a reasonable and checked default for -uroot-source
The current value of -uroot-source is now "."

For almost all users, this will be the right value. This is
preferable to the previous value, "", which was wrong for
ALL users, and, when not set, resulted in a confusing
error message.

Further, perform a simple test to see if the directory
cmds/core exists at -uroot-source, and print a hopefully
helpful error message if it does not.

This will lessen another surprise that came with modules support.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>